requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
transformers>=4.33.0
diffusers>=0.21.0
scikit-learn>=1.3.0
pillow>=9.4.0
matplotlib>=3.7.0
seaborn>=0.12.2
numpy>=1.24.0